Update AsyncOperation to use a reference counted OVERLAPPED structure. so that we can copy the AsyncOperation and mantain the same memory address of the OVERLAPPED structure.

+++ Pulp/Core/win32/OsFileAsync.h
@@ -14,8 +14,8 @@ struct OsFileAsync
 	OsFileAsync(const char* path, IFileSys::fileModeFlags mode);
 	~OsFileAsync(void);
 
-	XOsFileAsyncOperation OsFileAsync::readAsync(void* pBuffer, uint32_t length, uint32_t position);
-	XOsFileAsyncOperation OsFileAsync::writeAsync(const void* pBuffer, uint32_t length, uint32_t position);
+	XOsFileAsyncOperation OsFileAsync::readAsync(void* pBuffer, uint32_t length, size_t position);
+	XOsFileAsyncOperation OsFileAsync::writeAsync(const void* pBuffer, uint32_t length, size_t position);
 
 	size_t tell(void) const;
 	size_t remainingBytes(void) const;
